+++ CHANGES.md
@@ -5,6 +5,7 @@
 - remove VOOT support
 - remove compression framing support 
 - remove tls-auth support
+- remove "multi instance" support
 
 ## 1.4.9 (2018-12-05)
 - fix bug where disabling a user would only disconnect the user from the first